requirements.txt
openai~=1.58.0
numpy~=2.2.0
requests~=2.32.3
pillow~=11.1.0
ultralytics~=8.3.25
python-dotenv~=1.0.1
torch~=2.6.0
opencv-python~=4.10.0.84
httpx~=0.28.1
pydantic~=2.10.5
fastapi~=0.115.7
aiohttp~=3.11.11
uvicorn~=0.34.0
pytz~=2025.1
transformers~=4.49.0
torchvision~=0.21.0